+++ packages/app/src/overlay/canvas/Pool/Header.tsx
@@ -4,10 +4,10 @@
 import { faArrowsRotate } from '@fortawesome/free-solid-svg-icons'
 import { Polkicon } from '@w3ux/react-polkicon'
 import { determinePoolDisplay } from 'contexts/Pools/util'
+import { PageTabs } from 'library/PageTabs'
 import { useTranslation } from 'react-i18next'
 import type { PoolState } from 'types'
 import { ButtonPrimaryInvert } from 'ui-buttons'
-import { Page } from 'ui-core/base'
 import { AccountTitle, Head, HeadTags } from 'ui-core/canvas'
 import { CloseCanvas } from 'ui-overlay'
 import type { HeaderProps } from './types'
@@ -103,8 +103,7 @@ export const Header = ({
             </HeadTags>
           </div>
         </div>
-
-        <Page.TitleTabs
+        <PageTabs
           sticky={false}
           tabs={[
             {
