+++ packages/app/src/pages/Validators/index.tsx
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { useFavoriteValidators } from 'contexts/Validators/FavoriteValidators'
+import { PageTabs } from 'library/PageTabs'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Page } from 'ui-core/base'
@@ -23,22 +24,23 @@ export const ValidatorsInner = () => {
 
   return (
     <>
-      <Page.Title
-        title={t('validators')}
-        tabs={[
-          {
-            title: t('allValidators'),
-            active: activeTab === 0,
-            onClick: () => setActiveTab(0),
-          },
-          {
-            title: t('favorites'),
-            active: activeTab === 1,
-            onClick: () => setActiveTab(1),
-            badge: String(favorites.length),
-          },
-        ]}
-      />
+      <Page.Title title={t('validators')}>
+        <PageTabs
+          tabs={[
+            {
+              title: t('allValidators'),
+              active: activeTab === 0,
+              onClick: () => setActiveTab(0),
+            },
+            {
+              title: t('favorites'),
+              active: activeTab === 1,
+              onClick: () => setActiveTab(1),
+              badge: String(favorites.length),
+            },
+          ]}
+        />
+      </Page.Title>
       {activeTab === 0 && <AllValidators />}
       {activeTab === 1 && <ValidatorFavorites />}
     </>
