+++ packages/types/src/components.ts
@@ -11,3 +11,21 @@ export interface ComponentBase {
 export type ComponentBaseWithClassName = ComponentBase & {
   className?: string
 }
+
+export type PageTitleProps = ComponentBase & {
+  tabs?: PageTitleTabProps[]
+  colorSecondary?: boolean
+  tabClassName?: string
+  inline?: boolean
+  title?: string
+  sticky?: boolean
+}
+
+export interface PageTitleTabProps {
+  sticky?: boolean
+  title: string
+  active: boolean
+  onClick: () => void
+  badge?: string | number
+  disabled?: boolean
+}
