+++ packages/ui-core/tsconfig.json
@@ -16,5 +16,5 @@
     "jsx": "react-jsx",
     "noEmit": true
   },
-  "include": ["src", "../../env.d.ts"]
+  "include": ["src", "../../env.d.ts", "../app/src/library/PageTabs"]
 }
